Extract shared CloudFront /www proxy function sources
Deduplicate allowlist, path rewrite, media-request, and error-response
CloudFront Functions used by public-www and training stacks. Add unit test
for allowlist paths and wire into test:infra.
