+++ backend/infrastructure/package.json
@@ -9,7 +9,7 @@
     "cdk": "cdk",
     "synth": "cdk synth",
     "deploy": "cdk deploy",
-    "test:infra": "ts-node --prefer-ts-exts scripts/assert-existing-db.ts && ts-node --prefer-ts-exts test/public-www-stack.test.ts && ts-node --prefer-ts-exts test/training-stack.test.ts && ts-node --prefer-ts-exts test/api-stack.test.ts && ts-node --prefer-ts-exts test/admin-env-size.test.ts && ts-node --prefer-ts-exts test/public-www-config-secret.test.ts",
+    "test:infra": "ts-node --prefer-ts-exts scripts/assert-existing-db.ts && ts-node --prefer-ts-exts test/cloudfront-www-proxy-functions.test.ts && ts-node --prefer-ts-exts test/public-www-stack.test.ts && ts-node --prefer-ts-exts test/training-stack.test.ts && ts-node --prefer-ts-exts test/api-stack.test.ts && ts-node --prefer-ts-exts test/admin-env-size.test.ts && ts-node --prefer-ts-exts test/public-www-config-secret.test.ts",
     "test:admin-web-stack": "ts-node --prefer-ts-exts test/admin-web-stack.test.ts",
     "postinstall": "node scripts/patch-bundled-aws-cdk-lib-deps.js"
   },
